Cache message in RAM. Protect against overrun.
Load the message into RAM on startup. Use the same load routine after
a new message is defined in the serial handler.

Add a stub for a future demo routine, which will be called if either
the message length is zero, or it is 255 (which is what EEPROM is set
to when erased).

In order to detect uninitialized EEPROM, our max message length is
limited to 254 characters.