add default messages for different badge types
read the last byte of eeprom to get the badge type and then use it to
look up a default message for that type of badge, if no message is
defined.

Also, only show the demo loop on startup if there is a message or
badge type defined. Otherwise, show the demo loop endlessly.