update serial programming protocol
update serial programming protocol to be more resistant to noise from
handling.

To update, send 9 garbage characters to make certain you are in the base
mode. space or newline work well for this.

Then:
To update the message:
1. Send three 0x03 characters
2. Send the message length
3. Send the CRC32 of the message (big-endian)
4. Send the CRC xor message length (on each byte)
5. Send the message

To change badge type default message:
1. send three 0x01 characters
2. send the badge type number three times (in ascii)
   0 = Speaker
   1 = Staff
   2 = GrrCON
   3 = VIP

The default message is only displayed if there is no other message
programmed, or if the programmed message has been corrupted (which
is not likely to happen).